Process input directory on startup
Summary: This patch ensures that all files present in the input directory is processed on daemon startup.

Test Plan: Start the daemon with a couple of md5sum files and their corresponding data files, and check the log output for evidence that they are being processed.

Reviewers: #ecdiss, martinp

Reviewed By: #ecdiss, martinp

Subscribers: martinp

Projects: #ecdiss

Maniphest Tasks: T1844

Differential Revision: https://phab.met.no/D1051