Remove md5sum from Productstatus output data and disable moves
This patch offloads md5sum checking and dataset moving. Effectively, it
means that datasets will be published regardless of their consistency.
The goal is to delegate this reponsibility to a machine with faster file
system access and more processing power.

Since data sets are not moved, it is neccessary to cache processed data
sets in order to know whether or not they have been posted to
Productstatus. This cache might grow very large, and so this
functionality is not added. The implication is that ECReceive will
search for a matching DataInstance resource on Productstatus, and if
found, it is not re-posted. This eliminates the problem of duplicate
messages if ECReceive is restarted, but might lead to issues if a file
is re-transferred. The solution is to mark the problematic file(s) as
"deleted" in Productstatus.

Finally, the program has been changed to run on Python3.